requirements.txt
# This is a list of Python packages required for the mini_project
# pip install -r requirements.txt

aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
anyio==4.3.0
archspec==0.1.2
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
click
colorama==0.4.6
Cython==3.0.9
databases==0.9.0
dataclasses-json==0.6.4
datasets==2.18.0
dill==0.3.8
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.1  
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0  
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.21.4
idna==3.7
Jinja2==3.1.3  
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.1
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchain-google-genai  
langchainhub==0.1.15
langserve==0.0.48
langsmith==0.1.23
MarkupSafe==2.1.5  
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector==2.2.9  
mysql-connector-python==8.3.0  
networkx==3.2.1
numpy==1.26.4  
openai==1.13.3
orjson==3.9.15
packaging>=23.2 
pandas==2.2.1
pillow==10.3.0  
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.0  
pydantic_core==2.18.1  
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
setuptools==68.2.2  
six==1.16.0
sniffio==1.3.1 
soupsieve==2.5
SQLAlchemy==2.0.28
sse-starlette==1.8.2
starlette>=0.37.2

